// src/app/models/patient.model.ts
export class Patient{

  id: number | undefined;
  opdnumber: string="";
  firstname: string="";
  lastname: string="";
  sex: string="";
  dateOfBirth: Date= new Date();
  housenumber: string="";
  village: string="";
  phonenumber: string="";
  pregnant: string="";
  createTime: Date= new Date();

}